# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Método Ponto B - RE/MAX SC API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - name: franchiseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts
      parameters:
        - name: franchiseId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/resolve:
    patch:
      operationId: resolveAlert
      tags: [alerts]
      summary: Resolve an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /help-requests:
    get:
      operationId: listHelpRequests
      tags: [helpRequests]
      summary: List help requests
      parameters:
        - name: franchiseId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of help requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HelpRequest"
    post:
      operationId: createHelpRequest
      tags: [helpRequests]
      summary: Create a help request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HelpRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpRequest"
  /help-requests/{id}:
    patch:
      operationId: updateHelpRequest
      tags: [helpRequests]
      summary: Update a help request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HelpRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpRequest"
...
  /progress-history:
    get:
      operationId: listProgressHistory
      tags: [dashboard]
      summary: List progress history
      parameters:
        - name: franchiseId
          in: query
          schema:
            type: integer
        - name: entityType
          in: query
          schema:
            type: string
        - name: entityId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Progress history list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressHistoryItem"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [master_admin, staff_regional, franqueado, responsavel_interno]
        franchiseId:
          type: ["integer", "null"]
        franchiseName:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [master_admin, staff_regional, franqueado, responsavel_interno]
        franchiseId:
          type: ["integer", "null"]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [master_admin, staff_regional, franqueado, responsavel_interno]
        franchiseId:
          type: ["integer", "null"]
        active:
          type: boolean
...
    Alert:
      type: object
      required: [id, type, severity, message, status, createdAt]
      properties:
        id:
          type: integer
        franchiseId:
          type: ["integer", "null"]
        franchiseName:
          type: ["string", "null"]
        goalId:
          type: ["integer", "null"]
        goalTitle:
          type: ["string", "null"]
        goalInitiativeId:
          type: ["integer", "null"]
        type:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        message:
          type: string
        status:
          type: string
          enum: [open, resolved]
        createdAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
    HelpRequest:
      type: object
      required: [id, franchiseId, userId, description, status, createdAt]
      properties:
        id:
          type: integer
        franchiseId:
          type: integer
        franchiseName:
          type: ["string", "null"]
        goalId:
          type: ["integer", "null"]
        goalTitle:
          type: ["string", "null"]
        goalInitiativeId:
          type: ["integer", "null"]
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        description:
          type: string
        status:
          type: string
          enum: [aberto, em_atendimento, resolvido]
        assignedTo:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        createdAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
    HelpRequestInput:
      type: object
      required: [franchiseId, description]
      properties:
        franchiseId:
          type: integer
        goalId:
          type: integer
        goalInitiativeId:
          type: integer
        description:
          type: string
    HelpRequestUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [aberto, em_atendimento, resolvido]
        assignedTo:
          type: integer
...
    ProgressHistoryItem:
      type: object
      required: [id, entityType, entityId, createdAt]
      properties:
        id:
          type: integer
        entityType:
          type: string
        entityId:
          type: integer
        franchiseId:
          type: ["integer", "null"]
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        previousProgress:
          type: ["number", "null"]
        newProgress:
          type: ["number", "null"]
        previousStatus:
          type: ["string", "null"]
        newStatus:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        createdAt:
          type: string