# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Método Ponto B - RE/MAX SC API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  franchiseId?: number | null;
  /** @nullable */
  franchiseName?: string | null;
  active: boolean;
  createdAt: string;
}